Add support for setting output in workflow tasks
